testing class SolarPanels and SolarPanelPro with fixes

+++ Java How to program/SolarSystemSizing/src/solarsystemsizing/SolarPanelPro.java
@@ -24,9 +24,9 @@ public class SolarPanelPro extends SolarPanels{
      * In this case, the power of the available solar panel will be used to size the system
      * and might not be the most optimum choice...
      */
-    public SolarPanelPro(float energy, float irradiance, int panelPower) {
+    public SolarPanelPro(double energy, double irradiance, int panelPower) {
         super(energy, irradiance);
-        this.panelPower = panelPower;        
+        this.panelPower = panelPower;       
     }
 
     /**
@@ -37,7 +37,7 @@ public class SolarPanelPro extends SolarPanels{
      */
     @Override
     public int getPanelNumber() {
-        float power = (float)panelPower;
+        double power = (double)panelPower;
         int panelNumber = Math.round(super.getEnergy()/(super.getIrradiance()*power));
         if(!super.isEven(panelNumber)){
             panelNumber++;
